requirements.txt
asn1crypto==0.23.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.4
idna==2.5
inflection==0.3.1
linecache2==1.0.0
more-itertools==4.0.1
ndg-httpsclient==0.4.3
numpy==1.13.1
pandas==0.20.3
patsy==0.4.1
psutil==5.3.1
pyasn1==0.4.2
pycparser==2.18
pykalman==0.9.5
pyOpenSSL==17.5.0
python-dateutil==2.6.1
pytz==2017.2
pyzmq==16.0.3
Quandl==3.2.1
requests==2.17.3
scipy==1.0.0
six==1.11.0
statsmodels==0.8.0
TA-Lib==0.4.10
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.21.1
zmq==0.0.0
